requirements.txt
pandas>=2.2
numpy>=1.26
torch>=2.2
pytest>=8.0
pytest-sugar>=0.9
pytest-cov>=5.0
scikit-learn>=1.4
pyarrow>=15
matplotlib.pyplot>=3.9
seaborn>=0.12
